+++ src/main/kotlin/moe/shizuki/korrent/bittorrent/client/BitTorrentClient.kt
@@ -1,13 +1,12 @@
 package moe.shizuki.korrent.bittorrent.client
 
-import moe.shizuki.korrent.bittorrent.config.PluginConfigManager
-import moe.shizuki.korrent.bittorrent.data.PluginDataManager
+import moe.shizuki.korrent.plugin.config.PluginConfigManager
+import moe.shizuki.korrent.plugin.data.PluginDataManager
 import moe.shizuki.korrent.bittorrent.model.BitTorrentClientInfo
 
 abstract class BitTorrentClient {
     abstract val clientInfo: BitTorrentClientInfo
 
     val pluginConfigManager: PluginConfigManager = PluginConfigManager()
-
     val pluginDataManager = PluginDataManager()
 }
