--- src/main/kotlin/moe/shizuki/korrent/bittorrent/config/PluginCommonConfig.kt
@@ -1,3 +1,0 @@
-package moe.shizuki.korrent.bittorrent.config
-
-class PluginCommonConfig
