+++ src/main/kotlin/moe/shizuki/korrent/plugin/config/PluginCommonConfig.kt
@@ -1,0 +1,3 @@
+package moe.shizuki.korrent.plugin.config
+
+class PluginCommonConfig
